Add config for split operation files
